Move contrib embedding operators to a separate file
Move the `contrib` submodule of `src/ops/embedding.rs` to
`src/ops/embedding/contrib.rs` and move the tests for the contrib
operators into a `tests` module in that file.
